requirements.txt
streamlit
pandas
pillow
python-dotenv
google-generativeai
pymongo
matplotlib